Update UIParticle.cs to change default scale to Vector3.one
Hello - We've had some missteps with the default scaling of the particle system. Our VFX designer designs the particles but doesn't realize the default scale on the script is 10, and so has to go and do a lot of work editing the coefficients in the system (multiplying them all by 10) to get the VFX prefab to be correctly scaled.

Propose: The default value for scaling is Vector3.one instead of 10f.